# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for publishing and managing analytics events",
        "title": "Event Analytics Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/events": {
            "post": {
                "description": "Publish a single analytics event to the queue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Publish a single event",
                "parameters": [
                    {
                        "description": "Event data",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PublishEventRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.PublishEventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/bulk": {
            "post": {
                "description": "Publish multiple analytics events in bulk to the queue",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Publish multiple events",
                "parameters": [
                    {
                        "description": "Bulk events data",
                        "name": "events",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.PublishEventsBulkRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/models.PublishBulkEventsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the service is running",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "type": "string",
                    "example": "event_name is required"
                }
            }
        },
        "models.PublishBulkEventsResponse": {
            "type": "object",
            "properties": {
                "accepted": {
                    "type": "integer",
                    "example": 5
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "validation error on event 3"
                    ]
                },
                "event_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "evt_1",
                        "evt_2",
                        "evt_3"
                    ]
                },
                "rejected": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "models.PublishEventRequest": {
            "type": "object",
            "required": [
                "channel",
                "event_name",
                "timestamp",
                "user_id"
            ],
            "properties": {
                "campaign_id": {
                    "type": "string",
                    "example": "cmp_987"
                },
                "channel": {
                    "type": "string",
                    "example": "web"
                },
                "event_name": {
                    "type": "string",
                    "example": "product_view"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "price": "129.99",
                        "product_id": "prod-789"
                    }
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "electronics",
                        "homepage",
                        "flash_sale"
                    ]
                },
                "timestamp": {
                    "type": "integer",
                    "example": 1723475612
                },
                "user_id": {
                    "type": "string",
                    "example": "user_123"
                }
            }
        },
        "models.PublishEventResponse": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "string",
                    "example": "evt_1a2b3c4d5e6f"
                },
                "status": {
                    "type": "string",
                    "example": "accepted"
                }
            }
        },
        "models.PublishEventsBulkRequest": {
            "type": "object",
            "required": [
                "events"
            ],
            "properties": {
                "events": {
                    "type": "array",
                    "maxItems": 1000,
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/models.PublishEventRequest"
                    }
                }
            }
        }
    }
}